requirements.txt
anyio==3.4.0
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
click==8.0.3
colorama==0.3.7
command-not-found==0.3
contextlib2==21.6.0
contextvars==2.4
cryptography==2.1.4
cupshelpers==1.0
dataclasses==0.8
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
fastapi==0.70.1
greenlet==0.4.12
h11==0.12.0
h5py==2.10.0
httplib2==0.9.2
httptools==0.3.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
msgpack==0.5.6
neovim==0.2.0
netifaces==0.10.4
ninja==1.10.2.3
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==8.4.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.9.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.6.1
python-debian==0.1.32
python-dotenv==0.19.2
pytz==2018.3
pyxdg==0.25
PyYAML==6.0
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sniffio==1.2.0
ssh-import-id==5.7
starlette==0.16.0
system-service==0.3
systemd-python==234
torch==1.10.1
torchaudio==0.10.0
torchvision==0.11.2
typing-extensions==4.0.1
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
uvicorn==0.16.0
uvloop==0.14.0
virtualenv==15.1.0
wadllib==1.3.2
watchgod==0.7
websockets==9.1
xkit==0.0.0
zipp==3.6.0
zope.interface==4.3.2
